Add support for vaccum command.
https://trello.com/c/slNgx9tX/6-fix-vacuum

The VACUUM command works by copying the contents of the database into a
temporary database file and then overwriting the original with the
contents of the temporary file.

Due to the nature of the vacuum command, we do not add the security_context column when the vacuum
command is fired.

Add cunit test for regression.